Add functionality to set onboarding visited flag in cache
This commit adds a new function `setIsOnBoardingVisited()` to the `OnboardingController` class that sets a flag in the cache indicating that the onboarding process has been completed. The flag is set using the `CacheHelper` class. Additionally, the `OnboardingNavigationBar` widget now calls the `navigateToRegisterView()` function when the "Skip" or "Start Now" buttons are tapped.